tcr_hla_pep_predictor/requirements.txt
torch>=1.8.0
numpy>=1.20.0
pandas>=1.2.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
networkx>=2.5.0
biopython>=1.78
pyyaml>=5.4.0
tqdm>=4.60.0
tensorboard>=2.4.0 